refactor: extract file-tree key handling into key_actions module
